fix(CASH_BALANCE): stop casting through unknown for Send and guard the nested Add button
The half sheet's Send button forced a ParsedSearchAsset into ParsedAddressAsset via
`as unknown as`, which TypeScript only allowed because the cast bypassed its structural
checks; the two types actually diverge (network, mainnet_address, id, native.price.amount),
so Send analytics silently reported an empty network. Uses the existing
parsedSearchAssetToParsedAddressAsset converter instead, via a new shared useNavigateToSend
hook that also replaces SendActionButton's duplicated iOS/Android dispatch.

Also stops the wallet row's Add button from double-firing: it sits inside the row's own
ButtonPressAnimation, so its press now calls stopPropagation before navigating, matching the
same nested-button guard RnbwFeatureCard's DismissButton already uses.
